requirements.txt
altair==5.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.4
Django==4.2.9
folium==0.15.1
gitdb==4.0.11
GitPython==3.1.41
haversine==2.8.0
idna==3.6
image==1.5.33
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
patsy==0.5.6
pillow==10.2.0
plotly==5.18.0
plotly-express==0.4.1
protobuf==4.25.2
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
statsmodels==0.14.1
streamlit==1.29.0
streamlit-folium==0.18.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
xyzservices==2023.10.1
zipp==3.17.0
zope.interface==6.1
